Fix various typos and mishaps after refactor
