+++ db.js
@@ -1,14 +1,14 @@
 var mongoose = require('mongoose');
-var album = require('./album');
-var image = require('./image');
+var album = require('./model/album');
+var image = require('./model/image');
 mongoose.connect('mongodb://localhost/gallery');
 var mongoconnection = mongoose.connection;
 mongoconnection.on('error', console.error.bind(console, 'Error connecting to DB!'));
 mongoconnection.on('open', function cb() {
 	console.log("Connected to DB");
 });
-var imageSchema = image.getSchema(mongoose);
-var albumSchema = album.getSchema(mongoose);
+var imageSchema = image.imageSchema(mongoose);
+var albumSchema = album.albumSchema(mongoose);
 
 imageSchema.methods = {
 }
@@ -16,7 +16,6 @@ imageSchema.methods = {
 albumSchema.method = {
 		imageCount: function() { return this.images.lenght; },
 		getPath: function() { return this.images.path; },
-		getName: function () { return this.images.albumName; }.
+		getName: function () { return this.images.albumName; },
 		getPrintName: function() { return this.images.printName; }
 };
-
