+++ model/image.js
@@ -1,9 +1,10 @@
 function imageSchema(mongoose) {
 	return mongoose.Schema({
-		id : {type String, required: 'id is required', unique: true},
+		id : {type: String, required: 'id is required', unique: true},
 		imageName: {type: String, required: 'imageName is required'},
 		imageFilename: {type : String},
 		imageThumbnail: {type : String},
-		exif: {type Object}
+		exif: {type: Object}
 	});
 }
+exports.imageSchema = imageSchema;
